Add files via upload
Clarity Techworks Sdn. Bhd.'s ICT Jobs speak offered a brief but comprehensive overview of the variety of roles available within their ICT sector. With their detailed descriptions of the roles they play, responsibilities, and tools, each speaker gave us a glimpse into the diverse field of software engineering.The focus on the balanced value of technical knowledge and soft skills was important. Soft skills like communication, problem-solving, and attention to detail were as important as technical skills in platforms like Android Studio, Google Play Control panel, and Bitbucket. This comprehensive approach highlighted the field's difficulty, where success relies not only on technical skill but also on strong teamwork, communication, and flexibility.As a participant, it encouraged me to think about my own skills and goals for my career. It made clear just how crucial it is to keep learning and growing in order to be successful in the rapidly changing field of software engineering, both in terms of soft and hard abilities. I felt motivated and inspired by the opportunity to continue improving my skills and looking for opportunities in this great sector.