Add files via upload
The talk on the architecture and job scope of UTM Digital gave participants a useful understanding of the institution's technological abilities and data management techniques. Watched live on Webex, the talk gave an in-depth overview of UTM's goal and vision for developing a data-driven internet.It was interesting to learn about the design and features of the UTM Data Center, including virtual servers, databases, security measures for physical servers, and network security protocols. It was reassuring and eye-opening to learn how many precautions were put in place to ensure data protection, accessibility, and integrity.
The talk additionally highlighted beneficial uses of digital technology in educational and professional contexts, including the acquisition of student IDs, access to the Microsoft 365 suite, and internet security measures. This knowledge enhances our understanding of digital technologies and provides us valuable skills for both our academic and professional development.If everything was thought of, the talk offered a helpful chance to explore the complex issues of data management and digital infrastructure in schools and colleges. I went away from it with a deeper awareness of the importance of sound data management processes and how they affect both academic and professional performance.