Add files via upload
The PC system assembly assignment was a realistic learning experience that required accuracy and dedication to detail. I installed the motherboard, connected the wires, mounted the components, and tested the functionality of the unit by carefully following the step-by-step guide. Dealing together with an expert to verify my work offered me even more confidence. My practical skills were improved, and I experienced an air of accomplishment when I finally understood the assembly procedure thanks to this task.