Add files via upload
It was a life-changing experience working on the Kanri subscription management app's design thinking project. It provided a framework of problem-solving with a focus on creativity and focused on users design. The project's stages—from developing awareness for users to testing and prototyping—offered a thorough understanding of software engineering difficulties. During the demonstration, working with the team and getting feedback confirmed the app's usefulness and efficacy. For me, it matched my objectives of improving technical competence and establishing a connection between theory and practice. The project improved my problem-solving abilities while demonstrating how design thinking can be applied practically in real-world situations. In the future, this experience will be a great starting point for new initiatives and ongoing development in the ever-expanding field of software engineering.